alias deprecated funcs by proxying calls, move opts.
