+++ go.mod
@@ -4,7 +4,7 @@ require (
 	github.com/ipfs/go-cid v0.0.1
 	github.com/ipfs/go-log v0.0.1
 	github.com/libp2p/go-libp2p-blankhost v0.0.1
-	github.com/libp2p/go-libp2p-core v0.0.0-20190416150442-54a1b70f07da
+	github.com/libp2p/go-libp2p-core v0.0.0-20190417121120-c9a30b9fbea8
 	github.com/libp2p/go-libp2p-swarm v0.0.2
 	github.com/multiformats/go-multihash v0.0.1
 )
