--- interface.go
@@ -1,12 +1,0 @@
-package discovery
-
-import moved "github.com/libp2p/go-libp2p-core/discovery"
-
-// Deprecated: use skel.Advertiser instead.
-type Advertiser = moved.Advertiser
-
-// Deprecated: use skel.Discoverer instead.
-type Discoverer = moved.Discoverer
-
-// Deprecated: use skel.Discovery instead.
-type Discovery = moved.Discovery
